retail_support_agent/requirements.txt
# Core AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.38
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0

# Vector Store & Embeddings
chromadb==0.4.24
faiss-cpu==1.8.0
sentence-transformers==2.7.0

# Advanced RAG
rank-bm25==0.2.2

# Evaluation
ragas==0.1.9
datasets==2.19.1

# Data Processing
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0

# Fine-tuning
transformers==4.41.1
torch==2.3.0
peft==0.11.1
accelerate==0.30.1

# UI & Visualization
streamlit==1.34.0
plotly==5.22.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
tqdm==4.66.4
